[WIP] Super rough first draft of pulling events from mongodb
This is not ready for review. But it accomplishes a unit test that
inserts a doc into mongo DB (using a mongodb started inside same docker
instance that uses in-memory storage), then returns all events in that
mongodb's events collection. Needs lots more work and cleanup! I'm
putting this out there now just to show that there is a modicum of
progress : )